iommu: ipu-iommu: fix -Wpointer-to-int-cast error
drivers/iommu/ipu-iommu.c:313:15: error: cast to smaller integer type 'int' from 'struct jqs_message_ipu_reg_values *' [-Werror,-Wpointer-to-int-cast]
                                __func__, (int)rsp);
                                          ^~~~~~~~
drivers/iommu/ipu-iommu.c:323:15: error: cast to smaller integer type 'int' from 'struct jqs_message_ipu_reg_values *' [-Werror,-Wpointer-to-int-cast]
                                __func__, (int)rsp);
                                          ^~~~~~~~

llvm warning when cast a pointer into int. The design purpose of this
casting is for return value of error code. Remove cast to avoid this
error from and printf with %d can represent the field as int type.

Bug: 171827315

+++ drivers/iommu/ipu-iommu.c
@@ -310,7 +310,7 @@ static void ipu_iommu_tlb_sync(void *priv)
 		MMU_SYNC, MMU_SYNC_SYNC_MASK, false /*is_read*/);
 	if (IS_ERR(rsp)) {
 		dev_err(iommu->dev, "%s Error (%d) writing to mmu sync register\n",
-				__func__, (int)rsp);
+				__func__, rsp);
 		return;
 	}
 
@@ -320,7 +320,7 @@ static void ipu_iommu_tlb_sync(void *priv)
 		if (IS_ERR(rsp)) {
 			dev_err(iommu->dev,
 				"%s Error (%d) reading status register\n",
-				__func__, (int)rsp);
+				__func__, rsp);
 			return;
 		}
 		if (!(rsp->regs[0].value & MMU_SYNC_SYNC_MASK))
